add user gauge rewards
